Scanner now takes &str (before: &[char])
I realized that taking a &str makes for a better interface: constructing
a &[char] from a String requires to allocate another vector.

I also changed the internals to use the Peekable<Chars> iterator of &str. This should mean that we don't have to reallocate and copy the entirety of the source code.

I also changed Token::Identifier and Token::StrLit to own their own
string instead of holding a slice of the input source code. Allocating a string for each use of a variable could have a big impact on performance.
However I've found that ruff takes the same approach:
https://github.com/astral-sh/ruff/blob/b3f3529499afb8b64355d9aa0ac5231f5edd3fb4/crates/ruff_python_parser/src/token.rs
So I guess this should be okay.